Improve MemoryChecker performance.

+++ src/app/memoryhandler.h
@@ -20,7 +20,7 @@ signals:
     void memoryFreed();
 public slots:
     void freeMemory(const MemoryState &state,
-                    const unsigned long long &minFreeBytes);
+                    const long &minFreeBytes);
     void memoryChecked(const int memKb, const int totMemKb);
 private:
     MemoryState mCurrentMemoryState = NORMAL_MEMORY_STATE;
